Relocate asset rename planner into identifier-case
Move plan/apply helpers from src/plugin/src/assets/rename.js into src/plugin/src/identifier-case/asset-renames.js so they sit beside the identifier-case planner they serve.
